Move download.sh to posix

+++ download.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/bin/sh
 #
 # download script for giti.sh
 
@@ -8,37 +8,37 @@ DOWNLOAD_URL="https://github.com/frncsdrk/${INSTALLABLE_NAME}/archive/master.tar
 EXTRACTED_DIR_NAME="${INSTALLABLE_NAME}-master"
 
 download() {
-  if [[ ! -d /tmp ]]; then
+  if [ ! -d /tmp ]; then
     mkdir /tmp
   fi
 
-  cd /tmp
+  cd "/tmp" || exit 1
 
-  if [[ ! -x "$(command -v wget)" ]]; then
+  if [ ! -x "$(command -v wget)" ]; then
     printf '%s\n' "Please install wget to download ${INSTALLABLE_NAME}"
   fi
 
-  if [[ ! -d "${INSTALLABLE_NAME}" ]]; then
+  if [ ! -d "${INSTALLABLE_NAME}" ]; then
     mkdir "${INSTALLABLE_NAME}"
   fi
 
-  local TAR_TARGET="${INSTALLABLE_NAME}.tar.gz"
+  TAR_TARGET="${INSTALLABLE_NAME}.tar.gz"
   wget -q "${DOWNLOAD_URL}" -O "${TAR_TARGET}"
   tar -xzf "${TAR_TARGET}"
   rm "${TAR_TARGET}"
 
   printf '%s\n' "Installing to ${INSTALL_DIRECTORY_PATH}"
 
-  if [[ ! -d "${INSTALL_DIRECTORY_PATH}" ]]; then
+  if [ ! -d "${INSTALL_DIRECTORY_PATH}" ]; then
     mkdir -p "${INSTALL_DIRECTORY_PATH}"
   fi
 
-  if [[ -d "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}" ]]; then
-    rm -r "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}"
+  if [ -d "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}" ]; then
+    rm -r "${INSTALL_DIRECTORY_PATH:?}/${INSTALLABLE_NAME:?}"
   fi
   mv "${EXTRACTED_DIR_NAME}" "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}"
 
-  cd "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}"
+  cd "${INSTALL_DIRECTORY_PATH}/${INSTALLABLE_NAME}" || exit 1
   ./setup.sh i
 
   printf '%s\n' "DONE"
